fix: error when inlineStyle is false and module is true
